refactor: move docker integration to its own folder
