+++ internal/docker/docker.go
@@ -1,4 +1,4 @@
-package internal
+package docker
 
 import (
 	"context"
