+++ internal/docker/docker_test.go
@@ -1,4 +1,4 @@
-package internal
+package docker
 
 import "testing"
